Refactor to use AssembleWalkthrough
